major refactor for socket prep
created separate route functions for auth, bug, user, answer, and
angular.  created auth and answer controller. created pipe folder in
angular-common.  debugged bug-view, user-profile and nav component.ts
files. changed user.ts to unpack array to clean up component.ts usage.
fixed faction name bug that was always defaulting to rebel alliance.
updated readme. added socket dependencies to package.json. removed
commented code and console.logs from numerous files.
