+++ backend/_helpers/logger.js
@@ -1,9 +1,8 @@
 const { createLogger, format, transports } = require('winston');
-const { combine, timestamp, label, prettyPrint } = format;
+const { combine, timestamp, prettyPrint } = format;
 
 module.exports = createLogger({
   transports: [
-    // new Winston.transports.Console(),
     new (transports.Console)({'timestamp':true})
   ],
   format: combine(
