+++ backend/routes/auth.routes.js
@@ -1,0 +1,8 @@
+const auth = require('../controllers/auth.controller');
+
+module.exports = function(app) {
+
+  app
+    .post('/authservice/validate/email/', auth.validateEmail)
+    .post('/authservice/authenticate', auth.login);
+}
