+++ data/test_query.sql
@@ -118,20 +118,20 @@ USE bug_hunter;
 --      GROUP BY b.bug_id
 --      ORDER BY b.bug_created DESC;
 
-      --  SELECT user_id,
-      --         faction_name,
-      --         first_name,
-      --         last_name,
-      --         email,
-      --         admin,
-      --         profile_img,
-      --         konami_unlock,
-      --         user_created,
-      --         user_updated
-      --    FROM users AS u
-      --    JOIN factions AS f
-      --      ON f.faction_id = u.faction_id
-      --   WHERE email = 'bugs@gmail.com';
+       SELECT user_id,
+              faction_name,
+              first_name,
+              last_name,
+              email,
+              admin,
+              profile_img,
+              konami_unlock,
+              user_created,
+              user_updated
+         FROM users AS u
+         JOIN factions AS f
+           ON f.faction_id = u.faction_id
+        WHERE email = 'pbreeden@codingdojo.com';
 
 
 --        SELECT
