+++ package.json
@@ -18,6 +18,7 @@
     "jsonwebtoken": "^8.4.0",
     "mysql": "^2.16.0",
     "path": "^0.12.7",
+    "socket.io": "^2.2.0",
     "winston": "^3.1.0"
   }
 }
