+++ public/package.json
@@ -29,6 +29,7 @@
     "core-js": "^2.5.4",
     "highlight": "^0.2.4",
     "ngx-quill": "^4.4.5",
+    "ngx-socket-io": "^2.1.1",
     "quill": "^1.3.6",
     "rxjs": "~6.3.3",
     "time-ago-pipe": "^1.3.2",
@@ -36,14 +37,16 @@
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.0",
+    "@angular-devkit/build-angular": "^0.13.6",
     "@angular/cli": "~7.1.4",
-    "@angular/compiler-cli": "~7.1.4",
+    "@angular/compiler-cli": "^7.2.10",
     "@angular/language-service": "~7.1.4",
-    "@types/node": "~8.9.4",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
     "codelyzer": "~4.5.0",
+    "concat": "^1.0.3",
+    "hoek": "^6.1.2",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.0.0",
@@ -51,15 +54,13 @@
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.1",
-    "concat": "^1.0.3",
-    "hoek": "^6.1.2",
     "node-minify": "^3.6.0",
     "node-sass": "^4.11.0",
     "npm-run-all": "^4.1.5",
-    "postcss-cli": "^6.1.0"
+    "postcss-cli": "^6.1.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.1.1"
   }
 }
