+++ public/src/app/app.module.ts
@@ -19,8 +19,8 @@ import { AboutComponent } from './components/util/about/about.component';
 import { AutofocusDirective } from './common/directives/autofocus.directive';
 import { UniqueEmailValidatorDirective } from './common/directives/unique-email-validator.directive';
 import { UsersProfileComponent } from './components/users/users-profile/users-profile.component';
-import { FilterPipe } from './common/filter.pipe';
-import { PrettyPrintPipe } from './common/prettyprint.pipe';
+import { FilterPipe } from './common/pipes/filter.pipe';
+import { PrettyPrintPipe } from './common/pipes/prettyprint.pipe';
 import { TimeAgoPipe } from 'time-ago-pipe';
 
 @NgModule({
