+++ public/src/app/common/pipes/filter.pipe.ts
