+++ public/src/app/common/pipes/prettyprint.pipe.ts
